requirements.txt
flask
flask-cors
groq
python-dotenv
fpdf2
PyMuPDF
gunicorn